fix(proxy): surface incomplete stream compactions
